+++ src/RequestForm.test.js
@@ -2,6 +2,8 @@ import React from 'react';
 import {
   render,
   screen,
+  within,
+  fireEvent,
 } from '@testing-library/react';
 
 import '../test/jest/__mock__';
@@ -10,11 +12,16 @@ import { Field } from 'redux-form';
 
 import {
   Checkbox,
+  TextField,
 } from '@folio/stripes/components';
 
 import RequestForm from './RequestForm';
+import TitleInformation from './components/TitleInformation';
+import ItemDetail from './ItemDetail';
 
-const mockedTlrSettings = { titleLevelRequestsFeatureEnabled: true };
+import { REQUEST_LEVEL_TYPES } from './constants';
+
+let mockedTlrSettings;
 
 jest.mock('./utils', () => ({
   ...jest.requireActual('./utils'),
@@ -23,76 +30,79 @@ jest.mock('./utils', () => ({
 }));
 jest.mock('@folio/stripes/form', () => () => jest.fn((component) => component));
 jest.mock('redux-form', () => ({
-  Field: jest.fn(() => null),
+  // eslint-disable-next-line jsx-a11y/click-events-have-key-events, jsx-a11y/no-static-element-interactions
+  Field: jest.fn(({ onChange, ...rest }) => <div onClick={onChange} {...rest} />),
   getFormValues: jest.fn((formName) => (state) => state[formName]),
 }));
 jest.mock('./PatronBlockModal', () => jest.fn(() => null));
 jest.mock('./CancelRequestDialog', () => jest.fn(() => null));
+jest.mock('./components/TitleInformation', () => jest.fn(() => null));
+jest.mock('./ItemDetail', () => jest.fn(() => null));
+jest.mock('./PositionLink', () => jest.fn(() => null));
 
 describe('RequestForm', () => {
   const testIds = {
     tlrCheckbox: 'tlrCheckbox',
+    instanceInfoSection: 'instanceInfoSection',
   };
   const labelIds = {
     tlrCheckbox: 'ui-requests.requests.createTitleLevelRequest',
     instanceInformation: 'ui-requests.instance.information',
+    enterButton:'ui-requests.enter',
   };
   const fieldCallOrder = {
     tlrCheckbox: 1,
-  };
-  const mockedRequestForm = {
-    createTitleLevelRequest: true,
-  };
-  const mockedRequest = {
-    id: null,
-  };
-  const mockedQuery = {
-    itemId: null,
-    itemBarcode: null,
-    instanceId: null,
+    instanceHrid: 2,
   };
   const mockedChangeFunction = jest.fn();
-  const defaultProps = {
-    stripes: {
-      connect: jest.fn((component) => component),
-      store: {
-        getState: jest.fn().mockReturnValue({
-          requestForm: mockedRequestForm,
-        }),
-      },
-    },
-    change: mockedChangeFunction,
-    handleSubmit: jest.fn(),
-    asyncValidate: jest.fn(),
-    findResource: jest.fn(() => new Promise((resolve) => resolve({}))),
-    request: mockedRequest,
-    initialValues: {},
-    location: {
-      pathname: 'pathname',
-    },
-    onCancel: jest.fn(),
-    parentMutator: {
-      proxy: {
-        reset: jest.fn(),
-        GET: jest.fn(),
-      },
-    },
-    query: mockedQuery,
-    onSubmit: jest.fn(),
-    parentResources: {
-      patronBlocks: {
-        records: [],
-      },
-    },
-  };
 
-  beforeEach(() => {
+  const renderComponent = (passedProps = {}) => {
+    const {
+      mockedRequestForm = {},
+      mockedRequest = {},
+      mockedQuery = {},
+    } = passedProps;
+
+    const props = {
+      stripes: {
+        connect: jest.fn((component) => component),
+        store: {
+          getState: jest.fn().mockReturnValue({
+            requestForm: mockedRequestForm,
+          }),
+        },
+      },
+      change: mockedChangeFunction,
+      handleSubmit: jest.fn(),
+      asyncValidate: jest.fn(),
+      findResource: jest.fn(() => new Promise((resolve) => resolve({}))),
+      request: mockedRequest || {},
+      initialValues: {},
+      location: {
+        pathname: 'pathname',
+      },
+      onCancel: jest.fn(),
+      parentMutator: {
+        proxy: {
+          reset: jest.fn(),
+          GET: jest.fn(),
+        },
+      },
+      query: mockedQuery || {},
+      onSubmit: jest.fn(),
+      parentResources: {
+        patronBlocks: {
+          records: [],
+        },
+      },
+    };
+
     render(
       <RequestForm
-        {...defaultProps}
+        {...props}
       />
     );
-  });
+  };
 
   afterEach(() => {
     Field.mockClear();
@@ -101,19 +111,21 @@ describe('RequestForm', () => {
 
   describe('when `TLR` in enabled', () => {
     beforeAll(() => {
-      mockedTlrSettings.titleLevelRequestsFeatureEnabled = true;
+      mockedTlrSettings = {
+        titleLevelRequestsFeatureEnabled: true,
+      };
     });
 
     describe('when `isEdit` is false', () => {
-      beforeAll(() => {
-        mockedRequest.id = null;
-      });
-
       it('should render `TLR` checkbox section', () => {
+        renderComponent();
+
         expect(screen.getByTestId(testIds.tlrCheckbox)).toBeVisible();
       });
 
-      it('should execute `TLR` Field with passed props', () => {
+      it('should execute `TLR`checkbox Field with passed props', () => {
+        renderComponent();
+
         const expectedResult = {
           name: 'createTitleLevelRequest',
           type: 'checkbox',
@@ -124,98 +136,206 @@ describe('RequestForm', () => {
 
         expect(Field).toHaveBeenNthCalledWith(
           fieldCallOrder.tlrCheckbox,
-          expectedResult,
+          expect.objectContaining(expectedResult),
           {},
         );
       });
 
-      describe('`TLR` checkbox handle on first render', () => {
-        afterAll(() => {
-          mockedQuery.itemId = null;
-          mockedQuery.itemBarcode = null;
-          mockedQuery.instanceId = null;
-        });
+      it('should reset instance and item info on TLR checkbox change', () => {
+        renderComponent();
 
+        fireEvent.click(screen.getByTestId(testIds.tlrCheckbox));
+
+        expect(mockedChangeFunction).toBeCalledWith('item.barcode', null);
+        expect(mockedChangeFunction).toBeCalledWith('instance.hrid', null);
+        expect(mockedChangeFunction).toBeCalledWith('instanceId', null);
+      });
+
+      describe('`TLR` checkbox handle on first render', () => {
         describe('when only `itemId` is passed in query', () => {
-          beforeAll(() => {
-            mockedQuery.itemId = 'itemId';
-            mockedQuery.itemBarcode = null;
-            mockedQuery.instanceId = null;
-          });
+          const mockedQuery = {
+            itemId: 'itemId',
+          };
 
           it('should set form `createTitleLevelRequest` value to false', () => {
+            renderComponent({
+              mockedQuery,
+            });
+
             expect(mockedChangeFunction).toHaveBeenCalledWith('createTitleLevelRequest', false);
           });
         });
 
         describe('when only `itemBarcode` is passed in query', () => {
-          beforeAll(() => {
-            mockedQuery.itemId = null;
-            mockedQuery.itemBarcode = 'itemBarcode';
-            mockedQuery.instanceId = null;
-          });
+          const mockedQuery = {
+            itemBarcode: 'itemBarcode',
+          };
 
           it('should set form `createTitleLevelRequest` value to false', () => {
+            renderComponent({
+              mockedQuery,
+            });
+
             expect(mockedChangeFunction).toHaveBeenCalledWith('createTitleLevelRequest', false);
           });
         });
 
         describe('when only `instanceId` is passed in query', () => {
-          beforeAll(() => {
-            mockedQuery.itemId = null;
-            mockedQuery.itemBarcode = null;
-            mockedQuery.instanceId = 'instanceId';
-          });
+          const mockedQuery = {
+            instanceId: 'instanceId',
+          };
 
           it('should set form `createTitleLevelRequest` value to true', () => {
+            renderComponent({
+              mockedQuery,
+            });
+
             expect(mockedChangeFunction).toHaveBeenCalledWith('createTitleLevelRequest', true);
           });
         });
       });
 
       describe('when `TLR` checkbox is checked', () => {
-        beforeAll(() => {
-          mockedRequestForm.createTitleLevelRequest = true;
-        });
+        const mockedRequestForm = {
+          createTitleLevelRequest: true,
+        };
 
         it('should render Accordion with `Instance` information', () => {
+          renderComponent({
+            mockedRequestForm,
+          });
+
           expect(screen.getByText(new RegExp(labelIds.instanceInformation))).toBeVisible();
+        });
+
+        it('should render search filed for instance', () => {
+          renderComponent({
+            mockedRequestForm,
+          });
+
+          const instanceSection = screen.getByTestId(testIds.instanceInfoSection);
+          const expectedResult = {
+            name: 'instance.hrid',
+            component: TextField,
+          };
+
+          expect(Field).toHaveBeenNthCalledWith(fieldCallOrder.instanceHrid, expect.objectContaining(expectedResult), {});
+          expect(within(instanceSection).getByText(labelIds.enterButton)).toBeVisible();
         });
       });
 
       describe('when `TLR` checkbox is unchecked', () => {
-        beforeAll(() => {
-          mockedRequestForm.createTitleLevelRequest = false;
-        });
+        const mockedRequestForm = {
+          createTitleLevelRequest: false,
+        };
 
         it('should not render Accordion with `Instance` information', () => {
+          renderComponent({
+            mockedRequestForm,
+          });
+
           expect(screen.queryByText(new RegExp(labelIds.instanceInformation))).not.toBeInTheDocument();
         });
       });
     });
 
     describe('when `isEdit` is true', () => {
-      beforeAll(() => {
-        mockedRequest.id = 'testId';
-      });
+      const mockedInstance = {
+        id: 'instanceId',
+        title: 'instanceTitle',
+        contributors: 'instanceContributors',
+        publication: 'instancePublication',
+        editions: 'instanceEditions',
+        identifiers: 'instanceIdentifiers',
+      };
+      const mockedRequest = {
+        instance : mockedInstance,
+        id : 'testId',
+        instanceId : 'instanceId',
+      };
 
       it('should not render `TLR` checkbox section', () => {
+        renderComponent({
+          mockedRequest,
+        });
+
         expect(screen.queryByTestId(testIds.tlrCheckbox)).not.toBeInTheDocument();
+      });
+
+      describe('when `Title level` request is editing', () => {
+        it('should execute `TitleInformation` with passed props', () => {
+          renderComponent({
+            mockedRequest: {
+              ...mockedRequest,
+              requestLevel: REQUEST_LEVEL_TYPES.TITLE,
+            },
+          });
+
+          const expectedResult = {
+            instanceId: mockedInstance.id,
+            titleLevelRequestsCount: mockedRequest.position,
+            title: mockedInstance.title,
+            contributors: mockedInstance.contributors,
+            publications: mockedInstance.publication,
+            editions: mockedInstance.editions,
+            identifiers: mockedInstance.identifiers,
+          };
+
+          expect(TitleInformation).toHaveBeenCalledWith(expectedResult, {});
+        });
+      });
+
+      describe('when `Item level` request is editing', () => {
+        const mockedItem = {
+          barcode: 'itemBarcode',
+        };
+
+
+        it('should execute `ItemDetail` with passed props', () => {
+          renderComponent({
+            mockedRequest: {
+              ...mockedRequest,
+              item: mockedItem,
+              requestLevel: REQUEST_LEVEL_TYPES.ITEM,
+            },
+          });
+
+          const expectedResult = {
+            item: {
+              id: mockedRequest.itemId,
+              instanceId: mockedRequest.instanceId,
+              ...mockedItem,
+              ...mockedInstance,
+            },
+            requestCount: mockedRequest.requestCount,
+          };
+
+          expect(ItemDetail).toHaveBeenCalledWith(expect.objectContaining(expectedResult), {});
+        });
       });
     });
   });
 
   describe('when `TLR` is disabled', () => {
     beforeAll(() => {
-      mockedTlrSettings.titleLevelRequestsFeatureEnabled = false;
-      mockedRequest.id = null;
+      mockedTlrSettings = {
+        titleLevelRequestsFeatureEnabled: false,
+      };
     });
 
     it('should not display `TLR` checkbox', () => {
+      render(
+        renderComponent()
+      );
+
       expect(screen.queryByTestId(testIds.tlrCheckbox)).not.toBeInTheDocument();
     });
 
     it('should set form `createTitleLevelRequest` value to false', () => {
+      render(
+        renderComponent()
+      );
+
       expect(mockedChangeFunction).toHaveBeenCalledWith('createTitleLevelRequest', false);
     });
   });