+++ src/utils.test.js
@@ -1,43 +1,24 @@
-import React from 'react';
-
 import {
   escape,
 } from 'lodash';
 import '../test/jest/__mock__';
 
 import {
-  // buildLocaleDateAndTime,
   buildTemplate,
-  // buildUrl,
-  // convertToSlipData,
   createUserHighlightBoxLink,
   duplicateRequest,
   escapeValue,
-  // formatNoteReferrerEntityData,
-  // getFullName,
-  // getPatronGroup,
-  // getRequestTypeOptions,
-  // hasNonRequestableStatus,
-  // isDelivery,
-  // isNotYetFilled,
-  // isPagedItem,
-  // isPageRequest,
-  // parseErrorMessage,
-  // toUserAddress,
-  // userHighlightBox,
   getTlrSettings,
   getRequestLevelValue,
+  getInstanceRequestTypeOptions,
+  getInstanceQueryString,
 } from './utils';
 
 import {
-  // requestTypesByItemStatus,
-  // requestTypeOptionMap,
   itemStatuses,
-  // fulfilmentTypeMap,
-  // requestStatuses,
   requestTypesMap,
-  // requestTypes,
   REQUEST_LEVEL_TYPES,
+  REQUEST_TYPES,
 } from './constants';
 
 describe('escapeValue', () => {
@@ -195,6 +176,72 @@ describe('getRequestLevelValue', () => {
 
   it('should return `Item` if false is passed', () => {
     expect(getRequestLevelValue(false)).toBe(REQUEST_LEVEL_TYPES.ITEM);
+  });
+});
+
+describe('getInstanceRequestTypeOptions', () => {
+  const missedItem = {
+    status: {
+      name: itemStatuses.MISSING,
+    },
+  };
+  const availableItem = {
+    status: {
+      name: itemStatuses.AVAILABLE,
+    },
+  };
+  const checkedOutItem = {
+    status: {
+      name: itemStatuses.CHECKED_OUT,
+    },
+  };
+
+  it('should return `Page` request type if at least one available item is present', () => {
+    const expectedResult = [
+      REQUEST_TYPES[requestTypesMap.PAGE],
+    ];
+
+    expect(getInstanceRequestTypeOptions([missedItem, availableItem, checkedOutItem])).toEqual(expectedResult);
+  });
+
+  it('should return `Hold` request type if only missing items is present', () => {
+    const expectedResult = [
+      REQUEST_TYPES[requestTypesMap.HOLD],
+    ];
+
+    expect(getInstanceRequestTypeOptions([missedItem])).toEqual(expectedResult);
+  });
+
+  it('should return `Hold` and `Recall` request types if no available items and not all of them are missing', () => {
+    const expectedResult = [
+      REQUEST_TYPES[requestTypesMap.HOLD],
+      REQUEST_TYPES[requestTypesMap.RECALL],
+    ];
+
+    expect(getInstanceRequestTypeOptions([missedItem, checkedOutItem])).toEqual(expectedResult);
+  });
+});
+
+describe('getInstanceQueryString', () => {
+  const hrid = 'hrid';
+  const id = 'instanceId';
+
+  it('should return correct query string if both values are passed', () => {
+    const expectedResult = `("hrid"=="${hrid}" or "id"=="${id}")`;
+
+    expect(getInstanceQueryString(hrid, id)).toBe(expectedResult);
+  });
+
+  it('should return correct query string if only `hrid` value is passed', () => {
+    const expectedResult = `("hrid"=="${hrid}" or "id"=="${hrid}")`;
+
+    expect(getInstanceQueryString(hrid)).toBe(expectedResult);
+  });
+
+  it('should return correct query string if only `id` value is passed', () => {
+    const expectedResult = `("hrid"=="${id}" or "id"=="${id}")`;
+
+    expect(getInstanceQueryString(id)).toBe(expectedResult);
   });
 });