+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * Update request results page. Refs UIREQ-614.
 * Update request details pane. Refs UIREQ-613.
 * Closed TLR should not be able to be duplicated when TLR is disabled. Refs UIREQ-691.
+* Update form for posibility of `Title` level requests. Refs UIREQ-620.
 
 ## [6.0.0](https://github.com/folio-org/ui-requests/tree/v6.0.0) (2021-10-05)
 [Full Changelog](https://github.com/folio-org/ui-requests/compare/v5.1.0...v6.0.0)